Export bubble functionality into its own module
By the recent work the functionality to create bubbles in the UI has
grown. As the 'util' module they were living in so far is already quite
huge it was finally time to move them into their own dedicated module.
This also allowed to simplify the names of the factory functions as they
are now scoped by their module.
Furthermore this allowed to provide a module description for anyone new
who does not know what is meant with the term of a bubble or to bubble
content.

The hierarchy of the plugins modules is very flat so far. This makes it
a little hard to read through it and get a grasp of its implied
structure. As the UI has become a big part of this plugin it makes sense
to group them together in their own module directory. Though this is
only the "kick-off" and no other UI related modules have been created or
exported yet
